return PM3_SUCCESS on usart_writebuffer_sync, usb_write

+++ armsrc/apps.h
@@ -222,10 +222,10 @@ void iClass_Clone(uint8_t startblock, uint8_t endblock, uint8_t *data);
 void iClass_ReadCheck(uint8_t blockno, uint8_t keytype);
 
 // cmd.h
-int32_t reply_old(uint64_t cmd, uint64_t arg0, uint64_t arg1, uint64_t arg2, void *data, size_t len);
-int32_t reply_mix(uint64_t cmd, uint64_t arg0, uint64_t arg1, uint64_t arg2, void *data, size_t len);
-int32_t reply_ng(uint16_t cmd, int16_t status, uint8_t *data, size_t len);
-int16_t receive_ng(PacketCommandNG *rx);
+int reply_old(uint64_t cmd, uint64_t arg0, uint64_t arg1, uint64_t arg2, void *data, size_t len);
+int reply_mix(uint64_t cmd, uint64_t arg0, uint64_t arg1, uint64_t arg2, void *data, size_t len);
+int reply_ng(uint16_t cmd, int16_t status, uint8_t *data, size_t len);
+int receive_ng(PacketCommandNG *rx);
 
 // util.h
 void HfSniff(int, int);
